separate vector and non-vector testgen
